% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dlex_connect}
\alias{dlex_connect}
\title{Verbindung zur Datenbank herstellen (und Download falls nötig)}
\usage{
dlex_connect()
}
\description{
Verbindung zur Datenbank herstellen (und Download falls nötig)
}
