% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dlex_lookup}
\alias{dlex_lookup}
\title{Abruf lexikalischer Statistiken für eine Wortliste}
\usage{
dlex_lookup(word_list, db_con = NULL)
}
\arguments{
\item{word_list}{Ein Character-Vektor mit den Wörtern.}

\item{db_con}{Optional: Eine bestehende Verbindung.}
}
\value{
Ein Dataframe mit Statistiken.
}
\description{
Abruf lexikalischer Statistiken für eine Wortliste
}
