% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{dlex_regex}
\alias{dlex_regex}
\title{Suche nach Wörtern mit Regulären Ausdrücken (Regex)}
\usage{
dlex_regex(pattern, min_freq = 0, db_con = NULL)
}
\arguments{
\item{pattern}{Ein Regulärer Ausdruck (z.B. "^Ver.*ung$").}

\item{min_freq}{Minimale normalisierte Frequenz (Standard: 0).}

\item{db_con}{Optional: Eine bestehende Verbindung.}
}
\value{
Ein Dataframe mit den Treffern.
}
\description{
Suche nach Wörtern mit Regulären Ausdrücken (Regex)
}
